requirements.txt
backoff==2.2.1
datasets==3.6.0
huggingface_hub==0.32.2
openai==1.95.1
pydantic==2.11.7
pydantic_core==2.35.2
python-dotenv==1.1.1
scipy==1.16.0
tabulate==0.9.0
tqdm==4.66.4
vllm==0.9.0
